Created a class file to implement TableDataReader
